Set output-base-uri on p:xslt whenever needed
to ensure the output gets the base URI of the input, because we should
not rely on XMLCalabash to do this for us. The XProc specification
states that in absence of the output-base-uri option, the base URI of
the output document is the base URI of the first input document, but
it leaves open for interpretation what should happen with the
document's elements. If we use output-base-uri, we can be sure that
the document's elements get the base URI we expect (this option is
passed on straight to Saxon).